+++ CMakeLists.txt
@@ -7,9 +7,6 @@ find_package(Boost REQUIRED COMPONENTS system log)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY build)
 
 add_executable(hpcore
-    src/sock/socket_client.cpp
-    src/sock/socket_server.cpp
-    src/sock/socket_session.cpp
     src/p2p/peer_session_handler.cpp
     src/p2p/p2p.cpp
     src/util.cpp
