+++ src/sock/socket_session.hpp
@@ -6,6 +6,9 @@
 #include <bitset>
 #include <boost/asio.hpp>
 #include <boost/beast.hpp>
+#include <boost/beast/core.hpp>
+#include <boost/beast/websocket.hpp>
+#include "../util.hpp"
 #include "socket_session_handler.hpp"
 
 namespace beast = boost::beast;
@@ -14,7 +17,7 @@ namespace websocket = boost::beast::websocket;
 namespace http = boost::beast::http;
 
 using tcp = net::ip::tcp;
-using error = boost::system::error_code;
+using error_code = boost::system::error_code;
 
 namespace sock
 {
@@ -46,17 +49,17 @@ class socket_session : public std::enable_shared_from_this<socket_session<T>>
     beast::flat_buffer buffer_;               // used to store incoming messages
     websocket::stream<beast::tcp_stream> ws_; // websocket stream used send an recieve messages
     std::vector<T> queue_;                    // used to store messages temporarily until it is sent to the relevant party
-    socket_session_handler<T> &sess_handler_;    // handler passed to gain access to websocket events
+    socket_session_handler<T> &sess_handler_; // handler passed to gain access to websocket events
 
-    void fail(error ec, char const *what);
+    void fail(error_code ec, char const *what);
 
-    void on_accept(error ec);
+    void on_accept(error_code ec);
 
-    void on_read(error ec, std::size_t bytes_transferred);
+    void on_read(error_code ec, std::size_t bytes_transferred);
 
-    void on_write(error ec, std::size_t bytes_transferred);
+    void on_write(error_code ec, std::size_t bytes_transferred);
 
-    void on_close(error ec, std::int8_t type);
+    void on_close(error_code ec, std::int8_t type);
 
 public:
     socket_session(websocket::stream<beast::tcp_stream> &websocket, socket_session_handler<T> &sess_handler);
@@ -82,7 +85,7 @@ public:
     std::bitset<8> flags_;
 
     void server_run(const std::string &&address, const std::string &&port);
-    void client_run(const std::string &&address, const std::string &&port, error ec);
+    void client_run(const std::string &&address, const std::string &&port, error_code ec);
 
     void send(T msg);
 
@@ -91,5 +94,222 @@ public:
 
     void close();
 };
+
+template <class T>
+socket_session<T>::socket_session(websocket::stream<beast::tcp_stream> &websocket, socket_session_handler<T> &sess_handler)
+    : ws_(std::move(websocket)), sess_handler_(sess_handler)
+{
+    ws_.binary(true);
+}
+
+template <class T>
+socket_session<T>::~socket_session()
+{
+    sess_handler_.on_close(this);
+}
+
+//port and address will be used to identify from which client the message recieved in the handler
+template <class T>
+void socket_session<T>::server_run(const std::string &&address, const std::string &&port)
+{
+    port_ = port;
+    address_ = address;
+
+    //Set this flag to identify whether this socket session created when node acts as a server
+    flags_.set(util::SESSION_FLAG::INBOUND);
+
+    // Accept the websocket handshake
+    ws_.async_accept(
+        [sp = this->shared_from_this()](
+            error_code ec) {
+            sp->on_accept(ec);
+        });
+}
+
+//port and address will be used to identify from which server the message recieved in the handler
+template <class T>
+void socket_session<T>::client_run(const std::string &&address, const std::string &&port, error_code ec)
+{
+    port_ = port;
+    address_ = address;
+
+    if (ec)
+        return fail(ec, "handshake");
+
+    sess_handler_.on_connect(this);
+
+    ws_.async_read(
+        buffer_,
+        [sp = this->shared_from_this()](
+            error_code ec, std::size_t bytes) {
+            sp->on_read(ec, bytes);
+        });
+}
+
+/**
+ * Executes on error
+*/
+template <class T>
+void socket_session<T>::fail(error_code ec, char const *what)
+{
+    // LOG_ERR << what << ": " << ec.message();
+
+    // Don't report these
+    if (ec == net::error::operation_aborted ||
+        ec == websocket::error::closed)
+        return;
+}
+
+/**
+ * Executes on acceptance of new connection
+*/
+template <class T>
+void socket_session<T>::on_accept(error_code ec)
+{
+    // Handle the error, if any
+    if (ec)
+        return fail(ec, "accept");
+
+    sess_handler_.on_connect(this);
+
+    // Read a message
+    ws_.async_read(
+        buffer_,
+        [sp = this->shared_from_this()](
+            error_code ec, std::size_t bytes) {
+            sp->on_read(ec, bytes);
+        });
+}
+
+/*
+* Executes on completion of recieiving a new message
+*/
+template <class T>
+void socket_session<T>::on_read(error_code ec, std::size_t)
+{
+    //if something goes wrong when trying to read, socket connection will be closed and calling this to inform it to the handler
+    // read may get called when operation_aborted as well.
+    // We don't need to process read operation in that case.
+    if (ec == net::error::operation_aborted)
+        return;
+
+    // Handle the error, if any
+    if (ec)
+    {
+        // if something goes wrong when trying to read, socket connection will be closed and calling this to inform it to the handler
+        on_close(ec, 1);
+        return fail(ec, "read");
+    }
+
+    // Wrap the buffer data in a string_view and call session handler.
+    // We DO NOT transfer ownership of buffer data to the session handler. It should
+    // read and process the message and we will clear the buffer after its done with it.
+    const char *buffer_data = net::buffer_cast<const char *>(buffer_.data());
+    std::string_view message(buffer_data, buffer_.size());
+    sess_handler_.on_message(this, message);
+
+    // Clear the buffer
+    buffer_.consume(buffer_.size());
+
+    // Read another message
+    ws_.async_read(
+        buffer_,
+        [sp = this->shared_from_this()](
+            error_code ec, std::size_t bytes) {
+            sp->on_read(ec, bytes);
+        });
+}
+
+/*
+* Send message through an active websocket connection
+*/
+template <class T>
+void socket_session<T>::send(T msg)
+{
+    // Always add to queue
+    queue_.push_back(std::move(msg));
+
+    // Are we already writing?
+    if (queue_.size() > 1)
+        return;
+
+    std::string_view sv = queue_.front().buffer();
+
+    // We are not currently writing, so send this immediately
+    ws_.async_write(
+        // Project the outbound_message buffer from the queue front into the asio buffer.
+        net::buffer(sv.data(), sv.length()),
+        [sp = this->shared_from_this()](
+            error_code ec, std::size_t bytes) {
+            sp->on_write(ec, bytes);
+        });
+}
+
+/*
+* Executes on completion of write operation to a socket
+*/
+template <class T>
+void socket_session<T>::on_write(error_code ec, std::size_t)
+{
+    // Handle the error, if any
+    if (ec)
+        return fail(ec, "write");
+
+    // Remove the string from the queue
+    queue_.erase(queue_.begin());
+
+    // Send the next message if any
+    if (!queue_.empty())
+    {
+        std::string_view sv = queue_.front().buffer();
+        ws_.async_write(
+            net::buffer(sv.data(), sv.length()),
+            [sp = this->shared_from_this()](
+                error_code ec, std::size_t bytes) {
+                sp->on_write(ec, bytes);
+            });
+    }
+}
+
+/*
+* Close an active websocket connection gracefully
+*/
+template <class T>
+void socket_session<T>::close()
+{
+    // Close the WebSocket connection
+    ws_.async_close(websocket::close_code::normal,
+                    [sp = this->shared_from_this()](
+                        error_code ec) {
+                        sp->on_close(ec, 0);
+                    });
+}
+
+/*
+* Executes on completion of closing a socket connection
+*/
+//type will be used identify whether the error is due to failure in closing the web socket or transfer of another exception to this method
+template <class T>
+void socket_session<T>::on_close(error_code ec, std::int8_t type)
+{
+    // sess_handler_.on_close(this);
+
+    // if (type == 1)
+    //     return;
+
+    // if (ec)
+    //     return fail(ec, "close");
+}
+
+// When called, initializes the unique id string for this session.
+template <class T>
+void socket_session<T>::init_uniqueid()
+{
+    // Create a unique id for the session combining ip and port.
+    // We prepare this appended string here because we need to use it for finding elemends from the maps
+    // for validation purposes whenever a message is received.
+    uniqueid_.append(address_).append(":").append(port_);
+}
+
 } // namespace sock
 #endif
